Large scale changes to internals of the cli. Moved all click decorators to the respective cli module
